added a lot of comments and explanations for validators. Moved redundant tags and spansline tag selection to the schema
