+++ src/main/java/ca/nines/ise/validator/SectionCoverageValidator.java
@@ -1,7 +1,22 @@
+/*
+ * Copyright (C) 2016 Malcolm Moran <malcolm.moran@outlook.com>
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation version 2.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ */
 package ca.nines.ise.validator;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
 
 import ca.nines.ise.annotation.ErrorCode;
 import ca.nines.ise.dom.DOM;
@@ -12,6 +27,19 @@ import ca.nines.ise.node.Node;
 import ca.nines.ise.node.StartNode;
 import ca.nines.ise.node.TextNode;
 
+/**
+ * Checks to see if all text in the given DOM is inside at least one sectioning element.
+ * The rules are as follows,
+ *    If a PAGE tag exists in the document, all text must be within a PAGE tag.
+ *    For text outside of FRONTMATTER and BACKMATTER:
+ *      If an ACT or SCENE tag exist, all text must be within an ACT or SCENE tag.
+ *      If a DIV tag exists, all text must be within a DIV tag.
+ *    For text within FRONTMATTER or BACKMATTER:
+ *      All text within matter must be within a DIV tag.
+ *    If no sectioning exists, all text will be flagged with required.
+ *    
+ * @author Malcolm Moran <malcolm.moran@outlook.com>
+ */
 public class SectionCoverageValidator {
 
   ValidatorStack<StartNode> nodeStack;
