+++ src/main/java/ca/nines/ise/validator/ValidatorStack.java
@@ -1,6 +1,30 @@
+/*
+ * Copyright (C) 2016 Malcolm Moran <malcolm.moran@outlook.com>
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation version 2.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ */
 package ca.nines.ise.validator;
 
 import java.util.LinkedList;
+
+/**
+ * Stack to be used in validators to store nodes as they loop through a given DOM.
+ * ValidatorStack provides functions to compare and modify nodes in the stack via
+ * their names (.getName()).
+ * 
+ * @author Malcolm Moran <malcolm.moran@outlook.com>
+ */
 
 import ca.nines.ise.node.Node;
